+++ packages/ragbits-core/src/ragbits/core/sources/hf.py
@@ -7,7 +7,7 @@ from typing import ClassVar
 from ragbits.core.audit import trace, traceable
 from ragbits.core.sources.base import Source, get_local_storage_dir
 from ragbits.core.utils.decorators import requires_dependencies
-from ragbits.document_search.documents.exceptions import SourceConnectionError, SourceNotFoundError
+from ragbits.core.sources.exceptions import SourceConnectionError, SourceNotFoundError
 
 with suppress(ImportError):
     from datasets import load_dataset
