+++ packages/ragbits-document-search/tests/unit/test_git_source.py
@@ -7,7 +7,7 @@ import pytest
 
 from ragbits.core.sources.base import LOCAL_STORAGE_DIR_ENV
 from ragbits.core.sources.git import GitSource
-from ragbits.document_search.documents.exceptions import SourceNotFoundError
+from ragbits.core.sources.exceptions import SourceNotFoundError
 
 # Set the local storage directory for tests
 os.environ[LOCAL_STORAGE_DIR_ENV] = Path(__file__).parent.as_posix()
